devide Board component into container and component
